Reimplemented to use single shared VirtualAsyncEvent dispatcher.
This was done in order to get tVirtualAsyncThreadCoordination to work.
A good guiding principle for this seems to be to emulate AsyncEvent as
much as possible, and since AsyncEvent is based on FDs it is not
thread-local. This is now reflected in VirtualAsyncEvent as well and
enables threads to wait on a trigger on the same event coming from a
different thread.